+++ tests/snowpark/test_function.py
@@ -42,8 +42,78 @@ def test_deploy_function(
             handler='app.func1_handler'
             packages=()
             """
-        ),
+        ).strip(),
     ]
+
+
+@mock.patch("snowflake.connector.connect")
+@mock.patch("snowcli.cli.snowpark.commands.ObjectManager.describe")
+def test_deploy_function_with_external_access(
+    mock_describe,
+    mock_connector,
+    mock_ctx,
+    runner,
+    project_directory,
+):
+    mock_describe.side_effect = ProgrammingError("does not exist or not authorized")
+    ctx = mock_ctx()
+    mock_connector.return_value = ctx
+
+    with project_directory("snowpark_function_external_access") as project_dir:
+        result = runner.invoke(
+            [
+                "snowpark",
+                "deploy",
+            ],
+            catch_exceptions=False,
+        )
+
+    assert result.exit_code == 0, result.output
+    assert ctx.get_queries() == [
+        "create stage if not exists dev_deployment comment='deployments managed by snowcli'",
+        f"put file://{Path(project_dir).resolve()}/app.zip @dev_deployment/my_snowpark_project"
+        f" auto_compress=false parallel=4 overwrite=True",
+        dedent(
+            """\
+            create or replace function func1(a string, b variant)
+            returns string
+            language python
+            runtime_version=3.8
+            imports=('@dev_deployment/my_snowpark_project/app.zip')
+            handler='app.func1_handler'
+            packages=()
+            external_access_integrations=(external_1,external_2)
+            secrets=('cred'=cred_name,'other'=other_name)
+            """
+        ).strip(),
+    ]
+
+
+@mock.patch("snowflake.connector.connect")
+@mock.patch("snowcli.cli.snowpark.commands.ObjectManager.describe")
+def test_deploy_function_secrets_without_external_access(
+    mock_describe,
+    mock_connector,
+    mock_ctx,
+    runner,
+    project_directory,
+):
+    mock_describe.side_effect = ProgrammingError("does not exist or not authorized")
+    ctx = mock_ctx()
+    mock_connector.return_value = ctx
+
+    with project_directory("snowpark_function_secrets_without_external_access"):
+        result = runner.invoke(
+            [
+                "snowpark",
+                "deploy",
+            ],
+        )
+
+    assert result.exit_code == 1, result.output
+    assert result.output.__contains__(
+        "Can not provide secrets without external access integration"
+    )
 
 
 @mock.patch("snowflake.connector.connect")
@@ -131,7 +201,7 @@ def test_deploy_function_needs_update_because_packages_changes(
             handler='app.func1_handler'
             packages=('foo=1.2.3','bar>=3.0.0')
             """
-        ),
+        ).strip(),
     ]
 
 
@@ -182,7 +252,7 @@ def test_deploy_function_needs_update_because_handler_changes(
             handler='app.func1_handler'
             packages=('foo=1.2.3','bar>=3.0.0')
             """
-        ),
+        ).strip(),
     ]